# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: La Playita de Don Concho - Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Table detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TableDetail" }
  /tables/{id}/status:
    patch:
      operationId: updateTableStatus
      tags: [tables]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [free, occupied, waiting_payment, closed]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RestaurantTable" }
  /tables/{id}/open:
    post:
      operationId: openTable
      tags: [tables]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RestaurantTable" }
  /tables/{id}/transfer:
    post:
      operationId: transferTable
      tags: [tables]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetTableId]
              properties:
                targetTableId: { type: integer }
      responses:
        "200":
          description: Transferred
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
  /sales:
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSaleRequest" }
      responses:
        "200":
          description: Sale created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Sale" }
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaleDetail" }
...
  /analytics/sales:
    get:
      operationId: getSalesReport
      tags: [analytics]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesReport" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    RestaurantTable:
      type: object
      required: [id, areaId, areaName, number, status]
      properties:
        id: { type: integer }
        areaId: { type: integer }
        areaName: { type: string }
        number: { type: integer }
        status:
          type: string
          enum: [free, occupied, waiting_payment, closed]
        openedAt: { type: string, format: date-time, nullable: true }
        openedByName: { type: string, nullable: true }
        currentTotal: { type: number }
        currentOrderId: { type: integer, nullable: true }
    TableDetail:
      allOf:
        - $ref: "#/components/schemas/RestaurantTable"
        - type: object
          properties:
            order:
              nullable: true
              $ref: "#/components/schemas/OrderDetail"
...
    OrderItemDetail:
      type: object
      required: [id, productId, productName, quantity, unitPrice, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        notes: { type: string, nullable: true }
        lineTotal: { type: number }
    OrderDetail:
      type: object
      required: [id, tableId, waiterId, status, total, items, createdAt]
      properties:
        id: { type: integer }
        tableId: { type: integer }
        tableNumber: { type: integer }
        areaName: { type: string }
        waiterId: { type: integer }
        waiterName: { type: string }
        status: { type: string }
        total: { type: number }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemDetail" }
        createdAt: { type: string, format: date-time }
...
    CreateSaleRequest:
      type: object
      required: [orderId, payments]
      properties:
        orderId: { type: integer }
        payments:
          type: array
          items:
            type: object
            required: [method, amount]
            properties:
              method: { type: string, enum: [cash, card, transfer] }
              amount: { type: number }
    Sale:
      type: object
      required: [id, orderId, tableId, total, createdAt]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        tableId: { type: integer }
        tableNumber: { type: integer }
        cashierName: { type: string }
        waiterName: { type: string, nullable: true }
        subtotal: { type: number }
        total: { type: number }
        createdAt: { type: string, format: date-time }
        paymentMethods:
          type: array
          items: { type: string }
    SaleDetail:
      allOf:
        - $ref: "#/components/schemas/Sale"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  productName: { type: string }
                  quantity: { type: integer }
                  unitPrice: { type: number }
                  lineTotal: { type: number }
            payments:
              type: array
              items:
                type: object
                properties:
                  method: { type: string }
                  amount: { type: number }
    SalesReport:
      type: object
      required: [from, to, totalSales, transactionCount, sales]
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        totalSales: { type: number }
        transactionCount: { type: integer }
        averageTicket: { type: number }
        sales:
          type: array
          items: { $ref: "#/components/schemas/Sale" }
...